+++ packages/libs/extension/package.json
@@ -56,7 +56,6 @@
   "devDependencies": {
     "@types/jest": "^26.0.20",
     "@types/command-exists": "~1.2.0",
-    "@types/mocha": "^5.2.7",
     "@types/node": "~14.14.20",
     "@types/npm-package-arg": "^6.1.0",
     "@types/semver": "5.5.0",
