+++ packages/libs/extension/src/system-requirements/common.test.ts
@@ -1,47 +1,25 @@
-import { validateVersionRequirement, versionIsSatisfied } from "./common";
+import { getExecutablePath } from "./common";
 
-describe("System requirements common", () => {
-  describe("versionIsSatisfied", () => {
-    describe("when requirement is >=", () => {
-      it("is valid match exactly", () => {
-        expect(versionIsSatisfied("3.1.0", ">=3.1.0")).toBe(true);
-        expect(versionIsSatisfied("1.2.3", ">=1.2.3")).toBe(true);
-      });
-
-      it("is invalid if version is less", () => {
-        expect(versionIsSatisfied("3.0.0", ">=3.1.2")).toBe(false);
-        expect(versionIsSatisfied("3.1.0", ">=3.1.2")).toBe(false);
-      });
-
-      it("is valid if version is more", () => {
-        expect(versionIsSatisfied("3.2.0", ">=3.1.2")).toBe(true);
-        expect(versionIsSatisfied("3.1.3", ">=3.1.2")).toBe(true);
-      });
-
-      it("is valid if version contains tags and is more", () => {
-        expect(versionIsSatisfied("3.1.2-beta.19", ">=3.1.2")).toBe(true);
-        expect(versionIsSatisfied("3.2.0-rc.1", ">=3.1.2")).toBe(true);
-      });
-
-      it("is invalid if version contains tags but is less", () => {
-        expect(versionIsSatisfied("3.1.0-beta.19", ">=3.1.2")).toBe(false);
-        expect(versionIsSatisfied("3.0.0-rc.1", ">=3.1.2")).toBe(false);
-        expect(versionIsSatisfied("2.0.0-alpha.1", ">=3.1.2")).toBe(false);
-      });
+describe("System requirement common", () => {
+  describe("getExecutablePath", () => {
+    afterEach(() => {
+      delete process.env["MY_CUSTOM_TEST_PATH"];
     });
-  });
 
-  describe("validateVersionRequirement", () => {
-    it("is valid when version has beta flags", () => {
-      const resolution = { name: "test", command: "test" };
-      expect(validateVersionRequirement(resolution, "3.1.0-rc.2", { version: ">=3.1.0" })).toEqual({
-        command: "test",
-        name: "test",
-      });
-      expect(validateVersionRequirement(resolution, "3.1.0-beta.19", { version: ">=3.1.0" })).toEqual({
-        command: "test",
-        name: "test",
-      });
+    it("returns path described in env", () => {
+      process.env["MY_CUSTOM_TEST_PATH"] = "/path/to/exe";
+
+      expect(getExecutablePath({ environmentVariable: "MY_CUSTOM_TEST_PATH" })).toEqual("/path/to/exe");
+    });
+
+    it("returns undefined if environment is not set described in env", () => {
+      expect(getExecutablePath({ environmentVariable: "MY_CUSTOM_TEST_PATH" })).toBeUndefined();
+    });
+
+    it("returns undefined if requirement doesn't specify an environment variable name", () => {
+      process.env["MY_CUSTOM_TEST_PATH"] = "/path/to/exe";
+
+      expect(getExecutablePath({})).toBeUndefined();
     });
   });
 });
